Made HaversineDistance accept multiple types
By adding interfaces, HaversineDistance() can now accept any type. We have a type check to apply the
formula to the

HaversineFormula() was added to have a neutral function that can calculate the distance from
different data types. We made it public for users that want to calculate the Haversine distance
directly from floating points.